+++ src/components/navigation-bar.tsx
@@ -21,7 +21,9 @@ export function NavigationBar() {
 
   return (
     <div
-      className="z-100 fixed right-4 bottom-12 md:bottom-18 lg:bottom-1/2 lg:translate-y-1/2 flex flex-col items-center text-white max-w-min rounded-lg"
+      className={`z-100 fixed right-2 bottom-12 md:bottom-18 lg:bottom-1/2 lg:translate-y-1/2 flex flex-col items-center text-white max-w-min rounded-lg ${
+        isExpanded ? "opacity-100" : "opacity-50"
+      } transition-opacity duration-300`}
       onMouseEnter={() => setIsExpanded(true)}
       onMouseLeave={() => setIsExpanded(false)}
     >
@@ -38,7 +40,7 @@ export function NavigationBar() {
           ))}
         </div>
       ) : (
-        <div className="flex flex-col items-center p-4 space-y-1 bg-background">
+        <div className="flex flex-col items-center p-4 space-y-1 bg-background rounded-lg">
           {navItems.map((_, index) => (
             <div key={index} className="w-8 h-1 bg-primary/50 rounded-sm"></div>
           ))}